+++ CHANGES.md
@@ -11,6 +11,7 @@ Release Notes.
 * Add RocketMQ 5.x plugin
 * Fix the conflict between the logging kernel and the JDK threadpool plugin.
 * Fix the thread safety bug of finishing operation for the span named "SpringCloudGateway/sendRequest"
+* Add WebSphere Liberty 23.x plugin
 
 #### Documentation